refactor: reshape folder structure
