+++ Cargo.toml
@@ -18,3 +18,4 @@ open = "5.0.0"
 anyhow = "1.0"
 cliclack = "0.2.5"
 serde_json = "1.0.107"
+
