+++ src/commands/start.rs
@@ -5,8 +5,9 @@ use cliclack::{
 
 use crate::{
     cli::StartArgs,
-    connectors::{task::fetch_connector_task, utils::get_task_url_config_key},
     git::{GitBranch, GitConfig},
+    task_connectors::task_connector::TaskConnector,
+    utils::get_task_url_config_key,
 };
 
 pub async fn start_task(params: &StartArgs) -> anyhow::Result<()> {
@@ -16,7 +17,9 @@ pub async fn start_task(params: &StartArgs) -> anyhow::Result<()> {
 
     spinner.start("Fetching the task...");
 
-    let task_info = fetch_connector_task(params.link.as_str()).await?;
+    let connector = TaskConnector::from_url(&params.link.as_str())?;
+
+    let task_info = connector.fetch_task().await?;
 
     spinner.stop(format!("Found task with summary: {}", task_info.name));
