--- src/jira/mod.rs
@@ -1,3 +1,0 @@
-pub mod api;
-pub mod models;
-pub mod utils;
