+++ src/task_connectors/asana/mod.rs
@@ -1,4 +1,2 @@
 pub mod models;
-
-pub mod task;
 pub mod utils;
