+++ src/task_connectors/github/mod.rs
@@ -1,3 +1,2 @@
-pub mod api;
 pub mod models;
 pub mod utils;
